scaffold new PBR material component and unify defineMaterialPlugin internals with it
