+++ packages/engine/src/scene/functions/applyScreenshareToTexture.ts
@@ -5,8 +5,8 @@ import { defineQuery } from '@ir-engine/ecs/src/QueryFunctions'
 import { MeshComponent } from '@ir-engine/spatial/src/renderer/components/MeshComponent'
 
 import { Schema } from '@ir-engine/hyperflux'
+import { defineMaterialPlugin } from '@ir-engine/spatial'
 import { T } from '@ir-engine/spatial/src/schema/schemaFunctions'
-import { defineMaterialPlugin } from '../../material/defineMaterialPlugin'
 import { ScreenshareTargetComponent } from '../components/ScreenshareTargetComponent'
 import { fitTexture } from './fitTexture'
