+++ packages/spatial/src/renderer/materials/defineNodeMaterialPlugin.tsx
