+++ packages/spatial/src/renderer/materials/plugins/HolographicPlugin.ts
