+++ packages/spatial/src/renderer/materials/plugins/NoiseOffsetPlugin.test.ts
