+++ packages/spatial/src/renderer/materials/plugins/NoiseOffsetPlugin.ts
