+++ packages/spatial/src/renderer/materials/plugins/TransparencyDitheringComponent.test.tsx
