+++ packages/spatial/src/renderer/materials/plugins/ditherShaderChunk.ts
